Created Frontend
Made a basic frontend for two endpoints of my bookstore project. The user is greeted with a list of the top ten books in terms of copies sold, and has the option to add a book to the MySql database. The user is prompted with a form for entering information about the book, with a drop down list for genres & an option to cancel out to the list of books.